requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-Login==0.6.3

# Security
Werkzeug==2.3.7

# Development dependencies
python-dotenv==1.0.0

# Optional: For production deployment
gunicorn==21.2.0

# Optional: For testing
pytest==7.4.2
pytest-flask==1.2.0